# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio blog JSON API",
        "description": "Portfolio blog API documentation",
        "contact": {
            "email": "yseffk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/blog/item-attachments": {
            "get": {
                "tags": [
                    "Admin Content Attachment"
                ],
                "summary": "Get Content Item Attachments  whith pagination",
                "description": "Display a listing of the resource.",
                "operationId": "ce92a8297ce6c06713ef90b9a1011135",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "description": "Need pagination",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "blog_item_id",
                        "in": "query",
                        "description": "filtering by 'Content Item ID",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "filtering by source. enum('LINK','FILE') ",
                        "required": false,
                        "example": "LINK"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, created_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Content Attachment"
                ],
                "summary": "Create Content Item Attachment.",
                "operationId": "30d9de4fdc672922d49f8453ad8863c0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "blog_item_id",
                                    "source",
                                    "file_path"
                                ],
                                "properties": {
                                    "blog_item_id": {
                                        "description": "blogItem ID",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "source": {
                                        "description": "Item Attach Source.  enum('LINK','FILE')",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/item-attachments/{id}": {
            "get": {
                "tags": [
                    "Admin Content Attachment"
                ],
                "summary": "Show Content item Attachment.",
                "description": "Display the specified resource.",
                "operationId": "f9339ba254056965b861fe0b07d9909e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "itemAttachment ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Content Attachment"
                ],
                "summary": "Delete Content item Attachment by ID.",
                "description": "Display the specified resource.",
                "operationId": "c7426e92131138b97986d66c332fdb25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "itemAttachment ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/item-attachments/{id}?_method=PUT": {
            "post": {
                "tags": [
                    "Admin Content Attachment"
                ],
                "summary": "Update Content Item Attachment by ID.",
                "operationId": "a95f24726c6cf997473aaf2fcb0d771a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "blog_item_id",
                                    "source",
                                    "file_path"
                                ],
                                "properties": {
                                    "blog_item_id": {
                                        "description": "Content Item ID",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "source": {
                                        "description": "Item Attach Source.  enum('LINK','FILE')",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}